rafactor: rework file structure
