+++ src/components/App/AppRealization/index.js
@@ -1,0 +1,4 @@
+import AppCC from '@components/App/AppRealization/AppCC';
+import AppFC from '@components/App/AppRealization/AppFC';
+
+export { AppCC, AppFC };
