+++ src/components/Display/index.js
@@ -1,0 +1,4 @@
+import DisplayCC from '@components/Display/DisplayCC';
+import DisplayFC from '@components/Display/DisplayFC';
+
+export { DisplayCC, DisplayFC };
