+++ src/components/Header/HeaderFC/index.jsx
@@ -9,7 +9,7 @@ import {
   NavigationWrapper,
   Title,
 } from '@components/Header/styled';
-import { ModalCC, ModalFC } from '@components/Modal';
+import { ModalFC } from '@components/Modal';
 import routesLink from '@constants/links';
 
 const HeaderFC = () => {
@@ -45,45 +45,4 @@ const HeaderFC = () => {
   );
 };
 
-class HeaderCC extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isModalOpen: false,
-    };
-    this.links = (
-      <>
-        {routesLink.map(({ path, title }) => (
-          <NavLink to={path} data-cy={title} onClick={this.closeModal}>
-            {title}
-          </NavLink>
-        ))}
-      </>
-    );
-  }
-
-  closeModal = () => this.setState({ isModalOpen: false });
-
-  openModal = () => this.setState({ isModalOpen: true });
-
-  render() {
-    const { state, openModal, closeModal, links } = this;
-
-    return (
-      <HeaderWrapper>
-        <Title>Calculator App</Title>
-        <NavigationWrapper>{links}</NavigationWrapper>
-        <MobileNavigation onClick={openModal}>
-          <BiMenu />
-        </MobileNavigation>
-        {state.isModalOpen && (
-          <ModalCC handleClose={closeModal}>
-            <MobileNavigationWrapper>{links}</MobileNavigationWrapper>
-          </ModalCC>
-        )}
-      </HeaderWrapper>
-    );
-  }
-}
-
-export { HeaderCC, HeaderFC };
+export default HeaderFC;
