+++ src/components/Header/index.js
@@ -1,0 +1,4 @@
+import HeaderCC from '@components/Header/HeaderCC';
+import HeaderFC from '@components/Header/HeaderFC';
+
+export { HeaderCC, HeaderFC };
