+++ src/components/History/index.js
@@ -1,0 +1,4 @@
+import HistoryCC from '@components/History/HistoryCC';
+import HistoryFC from '@components/History/HistoryFC';
+
+export { HistoryCC, HistoryFC };
