+++ src/components/Modal/index.js
@@ -1,0 +1,4 @@
+import ModalCC from '@components/Modal/ModalCC';
+import ModalFC from '@components/Modal/ModalFC';
+
+export { ModalCC, ModalFC };
