+++ src/pages/Calculator/index.js
@@ -1,0 +1,6 @@
+import CalculatorCC from '@pages/Calculator/Calculator';
+import CalculatorFC from '@pages/Calculator/CalculatorFC';
+
+
+
+export { CalculatorCC, CalculatorFC };
